Making Authorized Requests

+++ frontend/src/components/workoutDetails.js
@@ -1,13 +1,24 @@
 import {useWorkoutsContext} from '../hooks/useWorkoutsContext'
+import { useAuthContext } from '../hooks/useAuthContext';
 //date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const WorkoutDetails = ({ workout }) => {
-    const { dispatch } = useWorkoutsContext();
+    const { dispatch } = useWorkoutsContext()
+    const {user} = useAuthContext()
 
     const handleClick = async () => {
+
+        if(!user){
+            return
+        }
+
         const response = await fetch('/api/workouts/' + workout._id, {
             method: 'DELETE',
+            headers : {
+                'Authorization' : `Bearer ${user.token}`
+
+            }
         });
         const json = await response.json();
